+++ packages/paste-website/src/components/tokens-list/index.tsx
@@ -50,7 +50,7 @@ interface TokensListProps {
 }
 
 const getTokensByTheme = (theme: ThemeVariants, props: TokensListProps): TokenCategory[] => {
-  const font = props.consoleTokens[0].node.tokens.find(ele => ele.categoryName === 'fonts');
+  const font = props.consoleTokens[0].node.tokens.find((ele) => ele.categoryName === 'fonts');
   if (font) {
     font.info = (
       <Callout>
@@ -65,7 +65,7 @@ const getTokensByTheme = (theme: ThemeVariants, props: TokensListProps): TokenCa
     );
   }
 
-  const fontSize = props.consoleTokens[0].node.tokens.find(ele => ele.categoryName === 'font-sizes');
+  const fontSize = props.consoleTokens[0].node.tokens.find((ele) => ele.categoryName === 'font-sizes');
   if (fontSize) {
     fontSize.info = (
       <Callout>
@@ -103,7 +103,7 @@ const getTokensByTheme = (theme: ThemeVariants, props: TokensListProps): TokenCa
 
 const collator = new Intl.Collator(undefined, {numeric: true, sensitivity: 'base'});
 
-export const TokensList: React.FC<TokensListProps> = props => {
+export const TokensList: React.FC<TokensListProps> = (props) => {
   const {theme} = useActiveSiteTheme();
   const [tokens, setTokens] = React.useState<TokenCategory[] | null>(getTokensByTheme(theme, props));
   const [filterString, setFilterString] = React.useState('');
@@ -112,13 +112,13 @@ export const TokensList: React.FC<TokensListProps> = props => {
     setTokens(() => {
       const newTokenCategories = getTokensByTheme(theme, props).map(
         (category): TokenCategory => {
-          const newTokens = category.tokens.filter(token => {
+          const newTokens = category.tokens.filter((token) => {
             return token.name.includes(filterString) || token.value.includes(filterString);
           });
           return {...category, tokens: newTokens};
         }
       );
-      const filteredCategories = newTokenCategories.filter(category => {
+      const filteredCategories = newTokenCategories.filter((category) => {
         return category.tokens.length > 0;
       });
       if (filteredCategories.length > 0) {
@@ -155,7 +155,7 @@ export const TokensList: React.FC<TokensListProps> = props => {
         />
       </Box>
       {tokens != null &&
-        tokens.map(cat => {
+        tokens.map((cat) => {
           return (
             <React.Fragment key={`catname${cat.categoryName}`}>
               <AnchoredHeading as="h2" variant="heading20">