+++ packages/paste-core/components/typography/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.96](https://github.com/twilio-labs/paste/compare/@twilio-paste/typography@1.0.95...@twilio-paste/typography@1.0.96) (2021-01-07)
+
+**Note:** Version bump only for package @twilio-paste/typography
+
+
+
+
+
 ## [1.0.95](https://github.com/twilio-labs/paste/compare/@twilio-paste/typography@1.0.94...@twilio-paste/typography@1.0.95) (2020-12-17)
 
 **Note:** Version bump only for package @twilio-paste/typography